refactor: reorganize app structure and localization
