+++ frontend/src/app/providers/createPinia.ts
@@ -1,0 +1,3 @@
+import { createPinia } from 'pinia'
+
+export const createAppPinia = () => createPinia()
