+++ frontend/src/features/payments/data/paymentLayout.ts
@@ -1,0 +1,51 @@
+import type { Locale } from '@/localization/store'
+import type { PaymentCurrency } from '@/features/payments/types'
+
+export type PaymentLayoutConfig = {
+  sectionOrder: PaymentCurrency[]
+  methodOrder?: Partial<Record<PaymentCurrency, string[]>>
+}
+
+const DEFAULT_CONFIG: PaymentLayoutConfig = {
+  sectionOrder: ['GLOBAL', 'KRW'],
+  methodOrder: {
+    GLOBAL: ['card', 'alipay', 'paypal'],
+    KRW: ['kakao', 'toss', 'transfer', 'naver'],
+  },
+}
+
+const LOCALE_CONFIGS: Partial<Record<Locale, PaymentLayoutConfig>> = {
+  ko: {
+    sectionOrder: ['KRW', 'GLOBAL'],
+  },
+  en: {
+    sectionOrder: ['GLOBAL', 'KRW'],
+    methodOrder: {
+      GLOBAL: ['card', 'paypal', 'alipay'],
+    },
+  },
+  zh: {
+    sectionOrder: ['GLOBAL', 'KRW'],
+    methodOrder: {
+      GLOBAL: ['alipay', 'card', 'paypal'],
+    },
+  },
+}
+
+export const getPaymentLayoutConfig = (locale: Locale): PaymentLayoutConfig => {
+  const localeConfig = LOCALE_CONFIGS[locale]
+
+  if (!localeConfig) {
+    return DEFAULT_CONFIG
+  }
+
+  const mergedMethodOrder = {
+    ...(DEFAULT_CONFIG.methodOrder ?? {}),
+    ...(localeConfig.methodOrder ?? {}),
+  }
+
+  return {
+    sectionOrder: localeConfig.sectionOrder ?? DEFAULT_CONFIG.sectionOrder,
+    methodOrder: Object.keys(mergedMethodOrder).length ? mergedMethodOrder : undefined,
+  }
+}
