+++ frontend/src/features/payments/workflows/actions/transfer.ts
@@ -1,0 +1,28 @@
+import type { PaymentActionContext, PaymentMethodAction } from '@/features/payments/workflows/types'
+
+const ensureTransferInfoReady = async (
+  context: PaymentActionContext,
+): Promise<boolean> => {
+  const loaded = await context.ensurePaymentInfoLoaded()
+
+  if (!loaded) {
+    return false
+  }
+
+  return true
+}
+
+const runTransferWorkflow = async (context: PaymentActionContext) => {
+  const ready = await ensureTransferInfoReady(context)
+
+  if (!ready) {
+    return
+  }
+
+  context.openTransferPopup()
+}
+
+export const createTransferAction = (context: PaymentActionContext): PaymentMethodAction => ({
+  handleSelection: () => runTransferWorkflow(context),
+  handleCurrencySelection: () => runTransferWorkflow(context),
+})
