+++ frontend/src/localization/messages/ko.json
@@ -1,0 +1,121 @@
+{
+  "hero": {
+    "kicker": "스티치몬 로드샵",
+    "title": "스티치몬 주문을 원하는 방식으로 결제해 보세요",
+    "badge": "공식 온라인 결제",
+    "description": "이곳은 스티치몬 로드샵 공식 온라인 결제 창구예요. 지금은 카카오송금으로 원화 결제를 도와드리고, 해외 고객을 위한 다양한 결제 수단도 순차적으로 준비하고 있어요."
+  },
+  "highlights": {
+    "instant": {
+      "title": "즉시 확인",
+      "description": "카카오송금으로 몇 초 만에 스티치몬 로드샵 주문을 결제하고 바로 확인하세요."
+    },
+    "guidance": {
+      "title": "친절한 안내",
+      "description": "각 결제 수단별 이용 방법을 단계별로 안내해 드려요."
+    },
+    "coverage": {
+      "title": "글로벌 준비",
+      "description": "해외 전자지갑과 카드 결제를 연동해 전 세계 스티치몬 팬도 편하게 결제하실 수 있도록 준비 중이에요."
+    }
+  },
+  "sections": {
+    "krw": {
+      "title": "KRW 결제",
+      "description": "스티치몬 로드샵 원화 주문을 바로 결제할 수 있는 수단이에요."
+    },
+    "global": {
+      "title": "글로벌 결제",
+      "description": "해외 스티치몬 고객을 위한 결제 옵션을 넓혀 가고 있어요."
+    }
+  },
+  "status": {
+    "available": "바로 이용 가능",
+    "comingSoon": "준비 중"
+  },
+  "card": {
+    "preparing": "곧 만나보실 수 있도록 마지막 손질을 하고 있어요.",
+    "selectedCurrency": "선택한 통화: {currency}",
+    "selectCurrencyPrompt": "진행하려면 통화를 선택해 주세요."
+  },
+  "footer": {
+    "message": "© {year} 스티치몬 로드샵. 공식 온라인 결제 센터입니다."
+  },
+  "language": {
+    "label": "언어"
+  },
+  "loading": {
+    "deepLink": "결제 앱을 여는 중이에요…"
+  },
+  "popups": {
+    "deepLink": {
+      "titles": {
+        "notMobile": "모바일에서만 이용할 수 있어요",
+        "notInstalled": "앱이 실행되지 않았나요?"
+      },
+      "confirms": {
+        "notMobile": "확인",
+        "notInstalled": "확인"
+      },
+      "providers": {
+        "kakao": {
+          "notMobile": "카카오송금 링크는 모바일 카카오톡 앱에서만 열려요. 휴대폰에서 다시 시도해 주세요.",
+          "notInstalled": "카카오톡으로 이동하지 않았다면 설치 여부를 확인한 후 다시 시도해 주세요."
+        },
+        "toss": {
+          "notMobile": "토스송금 링크는 모바일 토스 앱에서만 열려요. 휴대폰에서 다시 시도해 주세요.",
+          "notInstalled": "토스로 이동하지 않았다면 설치 여부를 확인한 후 다시 시도해 주세요."
+        }
+      }
+    }
+  },
+  "currencySelector": {
+    "title": "통화를 선택하세요",
+    "description": "{method}로 사용할 통화를 선택해 주세요.",
+    "cancel": "닫기"
+  },
+  "transferPopup": {
+    "title": "계좌이체 정보",
+    "description": "{amountWithCurrency} 금액을 아래 계좌 중 하나로 보내 주세요.",
+    "copyAll": "전체 정보 복사",
+    "copyNumber": "계좌번호만 복사",
+    "copyTooltip": "클립보드로 복사",
+    "copyAllButton": "이체 정보 복사",
+    "copiedAllButton": "클립보드에 복사됨",
+    "copied": "복사됨!",
+    "copiedNumber": "계좌번호만 복사했어요",
+    "close": "닫기"
+  },
+  "payment": {
+    "transfer": {
+      "name": "계좌이체",
+      "description": "원화 금액은 계좌이체로 보내 결제를 마무리해 주세요."
+    },
+    "kakao": {
+      "name": "카카오송금",
+      "description": "카카오톡으로 QR을 스캔해 빠르게 결제하세요.",
+      "cta": "카카오송금 열기"
+    },
+    "toss": {
+      "name": "토스송금",
+      "description": "토스송금으로 바로 결제하고 즉시 확인하세요.",
+      "cta": "토스송금 열기"
+    },
+    "naver": {
+      "name": "네이버페이",
+      "description": "현장에서도 네이버페이 포인트로 간편하게 결제하실 수 있도록 준비 중이에요."
+    },
+    "alipay": {
+      "name": "알리페이",
+      "description": "알리페이와 연동해 주요 글로벌 전자지갑을 연결하고, 여행 중에도 익숙한 지갑으로 결제하실 수 있도록 준비하고 있어요."
+    },
+    "paypal": {
+      "name": "페이팔",
+      "description": "글로벌 고객이 익숙한 PayPal 계정으로 손쉽게 결제하실 수 있도록 준비하고 있어요."
+    },
+    "card": {
+      "name": "신용카드",
+      "description": "Visa, Mastercard, UnionPay, Amex, JCB 등 주요 카드 결제를 곧 지원할 예정이에요."
+    }
+  }
+}
