+++ frontend/src/main.ts
@@ -1,11 +1,5 @@
 import './assets/main.css'
 
-import { createApp } from 'vue'
-import { createPinia } from 'pinia'
-import App from './App.vue'
+import { bootstrapApp } from '@/app/main'
 
-const app = createApp(App)
-
-app.use(createPinia())
-
-app.mount('#app')
+void bootstrapApp()
