+++ frontend/src/shared/components/LoadingOverlay.vue
