+++ frontend/src/shared/components/TooltipBubble.vue
