+++ frontend/src/shared/utils/clipboard.ts
