+++ frontend/src/shared/utils/device.ts
