+++ frontend/src/shared/utils/navigation.ts
